refactor: split actions and information display into separate components
Add a button to copy the url of a stream
